+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## 0.10.0
+
+July 9, 2026
+
+- Add provider wizard
+- Fix: logfile path
+- Fix: profile delete
+
 ## 0.9.1
 
 July 9, 2026